Streamlines install script to only perform needed actions

+++ setup/install.sh
@@ -1,7 +1,6 @@
 #!/usr/bin/bash
 
 set -e
-sudo rm -rf /var/www/html/svenska
 sudo apt install -y apache2 php php-curl
 script_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 post=${script_dir#*svenska}
@@ -27,18 +26,23 @@ mkdir -p $install_path/sounds/verb
 mkdir -p $install_path/sounds/adj
 mkdir -p $install_path/sounds/adv
 sudo chown -R www-data:www-data $install_path/sounds/*
-wget -O "$install_path/jquery-3.6.0.js" https://code.jquery.com/jquery-3.6.0.js
 
-wget -O "$install_path/jquery-ui-1.14.1.zip" https://jqueryui.com/resources/download/jquery-ui-1.14.1.zip
-eval "unzip $install_path/jquery-ui-1.14.1.zip -d $install_path"
+if [ ! -f $install_path/jquery-3.6.0.js ]; then
+    wget -O "$install_path/jquery-3.6.0.js" https://code.jquery.com/jquery-3.6.0.js
+fi
+
+if [ ! -d $install_path/jquery-ui-1.14.1 ]; then
+    wget -O "$install_path/jquery-ui-1.14.1.zip" https://jqueryui.com/resources/download/jquery-ui-1.14.1.zip
+    eval "unzip $install_path/jquery-ui-1.14.1.zip -d $install_path"
+fi
 
 # Local English to Swedish translation agent
-python3 -m venv $install_path/venv-libretranslate/
-source $install_path/venv-libretranslate/bin/activate
-pip install libretranslate
-sudo chown www-data:www-data -R $install_path/venv-libretranslate
-sudo cp $root_path/backend/libretranslate.service /etc/systemd/system/libretranslate.service
-sudo systemctl daemon-reload
-sudo systemctl restart libretranslate
+# python3 -m venv $install_path/venv-libretranslate/
+# source $install_path/venv-libretranslate/bin/activate
+# pip install libretranslate
+# sudo chown www-data:www-data -R $install_path/venv-libretranslate
+# sudo cp $root_path/backend/libretranslate.service /etc/systemd/system/libretranslate.service
+# sudo systemctl daemon-reload
+# sudo systemctl restart libretranslate
 
 sudo systemctl restart apache2
